requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
croniter==1.4.1
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.4
asyncio==3.4.3
